move cute reference to submission.py
