+++ problems/nvidia/fp16_gemm/reference.py
@@ -1,345 +1,15 @@
-import argparse
 import torch
-from typing import Tuple
-
-import cutlass
-import cutlass.cute as cute
-import cutlass.utils as utils
-import cutlass.torch as cutlass_torch
-import cutlass.pipeline as pipeline
-from cutlass.cute.nvgpu import cpasync, tcgen05
-import cutlass.utils.blackwell_helpers as sm100_utils
-from cutlass.cute.runtime import from_dlpack
-
-"""
-The first tutorial GEMM demonstrating a simple kernel implementation in CuTeDSL
-
-This dense GEMM kernel is implemented in just over 200 lines of code.
-With large tile sizes, it can achieve very high performance on 8k×8k×8k problem sizes.
-It can serve as a starting point to help users quickly experiment
-with optimizations for challenges that may arise with other problem sizes.
-
-To run this example:
-.. code-block:: bash
-
-    python examples/blackwell/tutorial_fp16_gemm_0.py  \
-      --mnk 8192,8192,8192                             \
-      --tolerance 1e-01
-
-Constraints for this example:
-* The problem size of m and n must be divisible by the tile size m & n (128, 256)
-"""
-
-io_dtype = cutlass.Float16
-acc_dtype = cutlass.Float32
-mma_inst_shape_mnk = (128, 256, 16)
-mma_tiler_mnk = (128, 256, 64)
-threads_per_cta = 128
-
-# Pipeline stage configuration
-ab_stages = 4
-acc_stage = 1
+from task import input_t, output_t
+from utils import make_match_reference
 
 
-@cute.struct
-class SharedStorage:
-    ab_mbar_ptr: cute.struct.MemRange[cutlass.Int64, ab_stages * 2]
-    acc_mbar_ptr: cute.struct.MemRange[cutlass.Int64, acc_stage * 2]
-    tmem_holding_buf: cutlass.Int32
-
-
-@cute.kernel
-def kernel(
-    tiled_mma: cute.TiledMma,
-    tma_atom_a: cute.CopyAtom,
-    mA_mkl: cute.Tensor,
-    tma_atom_b: cute.CopyAtom,
-    mB_nkl: cute.Tensor,
-    mC_mnl: cute.Tensor,
-    a_smem_layout: cute.ComposedLayout,
-    b_smem_layout: cute.ComposedLayout,
-):
-    # Current thread/warp/block coordinates
-    tidx, _, _ = cute.arch.thread_idx()
-    warp_idx = cute.arch.warp_idx()
-    warp_idx = cute.arch.make_warp_uniform(warp_idx)
-    bidx, bidy, _ = cute.arch.block_idx()
-    mma_coord_mnk = (bidx, bidy, None)
-
-    #
-    # 1. Prepare args
-    #
-
-    # Allocate SMEM
-    smem = cutlass.utils.SmemAllocator()
-    storage = smem.allocate(SharedStorage)
-    sA = smem.allocate_tensor(
-        element_type=io_dtype,
-        layout=a_smem_layout.outer,
-        byte_alignment=128,
-        swizzle=a_smem_layout.inner,
-    )
-    sB = smem.allocate_tensor(
-        element_type=io_dtype,
-        layout=b_smem_layout.outer,
-        byte_alignment=128,
-        swizzle=b_smem_layout.inner,
-    )
-
-    # Allocate all TMEM columns
-    # tmem_alloc_barrier = pipeline.NamedBarrier(
-    #     barrier_id=1,
-    #     num_threads=threads_per_cta,
-    # )
-    # tmem = utils.TmemAllocator(
-    #     storage.tmem_holding_buf,
-    #     barrier_for_retrieve=tmem_alloc_barrier,
-    # )
-    # num_tmem_cols = 512
-    # tmem.allocate(num_tmem_cols)
-
-    # Prefetch tma descriptor
-    if warp_idx == 0:
-        cpasync.prefetch_descriptor(tma_atom_a)
-        cpasync.prefetch_descriptor(tma_atom_b)
-
-    # Pipeline configuration
-    num_tma_copy_bytes = cute.size_in_bytes(
-        io_dtype, cute.select(a_smem_layout, mode=[0, 1, 2])
-    ) + cute.size_in_bytes(io_dtype, cute.select(b_smem_layout, mode=[0, 1, 2]))
-    ab_producer, ab_consumer = pipeline.PipelineTmaUmma.create(
-        num_stages=ab_stages,
-        producer_group=pipeline.CooperativeGroup(pipeline.Agent.Thread),
-        consumer_group=pipeline.CooperativeGroup(pipeline.Agent.Thread),
-        tx_count=num_tma_copy_bytes,
-        barrier_storage=storage.ab_mbar_ptr.data_ptr(),
-    ).make_participants()
-    acc_producer, acc_consumer = pipeline.PipelineUmmaAsync.create(
-        num_stages=acc_stage,
-        producer_group=pipeline.CooperativeGroup(pipeline.Agent.Thread),
-        consumer_group=pipeline.CooperativeGroup(
-            pipeline.Agent.Thread, threads_per_cta, threads_per_cta
-        ),
-        barrier_storage=storage.acc_mbar_ptr.data_ptr(),
-    ).make_participants()
-
-    # Partition tensors for MMA and make fragments
-    # (bM, bK, RestK)
-    gA = cute.local_tile(mA_mkl, mma_tiler_mnk, mma_coord_mnk, proj=(1, None, 1))
-    # (bN, bK, RestK)
-    gB = cute.local_tile(mB_nkl, mma_tiler_mnk, mma_coord_mnk, proj=(None, 1, 1))
-    # (bM, bN)
-    gC = cute.local_tile(mC_mnl, mma_tiler_mnk, mma_coord_mnk, proj=(1, 1, None))
-    thr_mma = tiled_mma.get_slice(0)
-    # (MMA, MMA_M, MMA_K)
-    tCgA = thr_mma.partition_A(gA)
-    # (MMA, MMA_N, MMA_K)
-    tCgB = thr_mma.partition_B(gB)
-    # (MMA, MMA_M, MMA_N)
-    tCgC = thr_mma.partition_C(gC)
-    # (MMA, MMA_M, MMA_K)
-    tCrA = tiled_mma.make_fragment_A(sA)
-    # (MMA, MMA_N, MMA_K)
-    tCrB = tiled_mma.make_fragment_B(sB)
-    # (MMA, MMA_M, MMA_N)
-    acc_shape = tiled_mma.partition_shape_C(mma_tiler_mnk[:2])
-    # (MMA, MMA_M, MMA_N)
-    tCtAcc = tiled_mma.make_fragment_C(acc_shape)
-    # Partition tensors for TMA; This requires the tensors partitioned for MMA
-    tAsA, tAgA = cute.nvgpu.cpasync.tma_partition(
-        tma_atom_a,
-        0,
-        cute.make_layout(1),
-        cute.group_modes(sA, 0, 3),
-        cute.group_modes(tCgA, 0, 3),
-    )
-    tBsB, tBgB = cute.nvgpu.cpasync.tma_partition(
-        tma_atom_b,
-        0,
-        cute.make_layout(1),
-        cute.group_modes(sB, 0, 3),
-        cute.group_modes(tCgB, 0, 3),
-    )
-
-    # CTA-wide sync before retrieving the pointer to the start of the allocated TMEM
-    # Only warp 0 does the allocation so we need to sync before retrieving the TMEM start address
-    if warp_idx == 0:
-        cute.arch.alloc_tmem(512, storage.tmem_holding_buf)
-    cute.arch.barrier()
-    tmem_ptr = cute.arch.retrieve_tmem_ptr(
-        acc_dtype,
-        alignment=16,
-        ptr_to_buffer_holding_addr=storage.tmem_holding_buf,
-    )
-
-    # Swap the pointer in tCtAcc
-    tCtAcc = cute.make_tensor(tmem_ptr, tCtAcc.layout)
-
-    subtile_cnt = 4
-    # (EpiTile)
-    epi_tiler = (
-        (cute.size(tCtAcc, mode=[0, 0]), cute.size(tCtAcc, mode=[0, 1]) // subtile_cnt),
-    )
-    # (EpiTile, NumTiles)
-    tCtAcc_epi = cute.zipped_divide(tCtAcc, epi_tiler)
-    # (EpiTile, NumTiles)
-    gC_epi = cute.zipped_divide(tCgC, epi_tiler)
-
-    # Every thread loads 32x128 bits
-    tmem_atom = cute.make_copy_atom(
-        tcgen05.Ld32x32bOp(tcgen05.Repetition.x64),
-        cutlass.Float32,
-    )
-    tmem_tiled_copy = tcgen05.make_tmem_copy(tmem_atom, tCtAcc_epi[None, 0])
-    tmem_thr_copy = tmem_tiled_copy.get_slice(tidx)
-
-    # (TmemCpy,NumTmemCpy,NumTiles)
-    tDtC = tmem_thr_copy.partition_S(tCtAcc_epi)
-    # (TmemCpy,NumTmemCpy,NumTiles)
-    tDgC = tmem_thr_copy.partition_D(gC_epi)
-
-    # (TmemCpy,NumTmemCpy)
-    tCrAcc = cute.make_fragment(tDgC[None, None, 0].shape, acc_dtype)
-    # (TmemCpy,NumTmemCpy)
-    tCrC = cute.make_fragment(tDgC[None, None, 0].shape, io_dtype)
-
-    #
-    # 2. Main loop
-    #
-    num_k_tiles = cute.size(gA, mode=[2])
-    if warp_idx == 0:
-        # Wait for a empty accumulator buffer
-        acc_empty = acc_producer.acquire_and_advance()
-        for k_tile_idx in cutlass.range(num_k_tiles, prefetch_stages=ab_stages - 2):
-            # Issue TMA loads
-            ab_empty = ab_producer.acquire_and_advance()
-            cute.copy(
-                tma_atom_a,
-                tAgA[(None, ab_empty.count)],
-                tAsA[(None, ab_empty.index)],
-                tma_bar_ptr=ab_empty.barrier,
-            )
-            cute.copy(
-                tma_atom_b,
-                tBgB[(None, ab_empty.count)],
-                tBsB[(None, ab_empty.index)],
-                tma_bar_ptr=ab_empty.barrier,
-            )
-
-            # Execute one K-block worth of MMA instructions
-            ab_full = ab_consumer.wait_and_advance()
-            num_k_blocks = cute.size(tCrA, mode=[2])
-            for k_block_idx in cutlass.range_constexpr(num_k_blocks):
-                k_block_coord = (None, None, k_block_idx, ab_full.index)
-                cute.gemm(
-                    tiled_mma,
-                    tCtAcc,
-                    tCrA[k_block_coord],
-                    tCrB[k_block_coord],
-                    tCtAcc,
-                )
-                tiled_mma.set(tcgen05.Field.ACCUMULATE, True)
-
-            # Signal that the A/B buffers have been consumed and are ready for the next load
-            ab_full.release()
-
-        # Signal that the accumulator is fully computed
-        acc_empty.commit()
-
-    #
-    # 3. Epilogue
-    #
-
-    # Release TMEM allocation lock
-    if warp_idx == 0:
-        cute.arch.relinquish_tmem_alloc_permit()
-
-    # Wait for the accumulator buffer to be full
-    acc_full = acc_consumer.wait_and_advance()
-
-    # TMEM -> RMEM -> GEMM
-    # Sub-tiling for better instruction-level parallelism
-    for i in cutlass.range(cute.size(tDtC, mode=[2])):
-        cute.copy(tmem_tiled_copy, tDtC[None, None, i], tCrAcc)
-        tCrC.store(tCrAcc.load().to(io_dtype))
-        cute.autovec_copy(tCrC, tDgC[None, None, i])
-
-    acc_full.release()
-
-    # Deallocate TMEM
-    cute.arch.barrier()
-    if warp_idx == 0:
-        cute.arch.dealloc_tmem(tmem_ptr, 512)
-
-
-@cute.jit
 def ref_kernel(
-    a: cute.Tensor,
-    b: cute.Tensor,
-    c: cute.Tensor,
-):
-    # Construct tiled MMA
-    op = tcgen05.MmaF16BF16Op(
-        io_dtype,
-        acc_dtype,
-        mma_inst_shape_mnk,
-        tcgen05.CtaGroup.ONE,
-        tcgen05.OperandSource.SMEM,
-        tcgen05.OperandMajorMode.K,
-        tcgen05.OperandMajorMode.K,
-    )
-    tiled_mma = cute.make_tiled_mma(op)
-
-    # Construct SMEM layouts for A and B
-    a_smem_layout = sm100_utils.make_smem_layout_a(
-        tiled_mma,
-        mma_tiler_mnk,
-        a.element_type,
-        ab_stages,
-    )
-    b_smem_layout = sm100_utils.make_smem_layout_b(
-        tiled_mma,
-        mma_tiler_mnk,
-        b.element_type,
-        ab_stages,
-    )
-    a_smem_layout_one_stage = cute.select(a_smem_layout, mode=[0, 1, 2])
-    b_smem_layout_one_stage = cute.select(b_smem_layout, mode=[0, 1, 2])
-
-    # Construct TMA load atoms
-    op = cute.nvgpu.cpasync.CopyBulkTensorTileG2SOp(tcgen05.CtaGroup.ONE)
-    a_tma_atom, a_tma_tensor = cute.nvgpu.make_tiled_tma_atom_A(
-        op,
-        a,
-        a_smem_layout_one_stage,
-        mma_tiler_mnk,
-        tiled_mma,
-        (1, 1, 1),
-    )
-    b_tma_atom, b_tma_tensor = cute.nvgpu.make_tiled_tma_atom_B(
-        op,
-        b,
-        b_smem_layout_one_stage,
-        mma_tiler_mnk,
-        tiled_mma,
-        (1, 1, 1),
-    )
-
-    # Launch the kernel
-    grid_shape = cute.ceil_div((*c.layout.shape, 1), mma_tiler_mnk[:2])
-    kernel(
-        tiled_mma,
-        a_tma_atom,
-        a_tma_tensor,
-        b_tma_atom,
-        b_tma_tensor,
-        c,
-        a_smem_layout,
-        b_smem_layout,
-    ).launch(
-        grid=grid_shape,
-        block=(threads_per_cta, 1, 1),
-    )
+    data: input_t,
+)->output_t:
+    a, b, c = data
+    # call torch matmul operation
+    c = torch.einsum("mk,nk->mn", a, b)
+    return c
 
 
 def generate_input(
@@ -347,49 +17,15 @@ def generate_input(
     n: int,
     k: int,
     seed: int,
-):
+) -> input_t:
     torch.manual_seed(seed)
 
-    # Make K-major tensors (torch tensors are row-major)
-    def make_tensors(mn, k, dtype):
-        shape = (mn, k)
-        return (
-            torch.empty(*shape, dtype=torch.int32)
-            .random_(-2, 2)
-            .to(dtype=dtype, device="cuda")
-        )
+    # Generate a, b and c tensors
+    a = torch.empty(m, k, dtype=torch.float16).random_(-2, 2).to(device="cuda")
+    b = torch.empty(n, k, dtype=torch.float16).random_(-2, 2).to(device="cuda")
+    c = torch.empty(m, n, dtype=torch.float16).random_(-2, 2).to(device="cuda")
 
-    a = make_tensors(m, k, cutlass_torch.dtype(io_dtype))
-    b = make_tensors(n, k, cutlass_torch.dtype(io_dtype))
-    c = make_tensors(m, n, cutlass_torch.dtype(io_dtype))
-    a_tensor = (
-        from_dlpack(a, assumed_align=32)
-        .mark_layout_dynamic(leading_dim=1)
-        .mark_compact_shape_dynamic(mode=1, divisibility=k)
-    )
-    b_tensor = (
-        from_dlpack(b, assumed_align=32)
-        .mark_layout_dynamic(leading_dim=1)
-        .mark_compact_shape_dynamic(mode=1, divisibility=k)
-    )
-    c_tensor = (
-        from_dlpack(c, assumed_align=32)
-        .mark_layout_dynamic(leading_dim=1)
-        .mark_compact_shape_dynamic(mode=1, divisibility=n)
-    )
-
-    # Compute reference result and verify
-    ref = (torch.einsum("mk,nk->mn", a.to(torch.float32), b.to(torch.float32))).cpu()
-
-    return a_tensor, b_tensor, c_tensor, c, ref
+    return (a, b, c)
 
 
-def check_implementation(ref, c_tensor):
-    torch.testing.assert_close(
-        c_tensor.cpu(), ref.to(cutlass_torch.dtype(io_dtype)), atol=1e-01, rtol=1e-05
-    )
-    print("PASS")
-
-a_tensor, b_tensor, c_tensor, c, ref = generate_input(m=1024, n=1024, k=1024, seed=1111)
-ref_kernel(a_tensor, b_tensor, c_tensor)
-check_implementation(ref, c)
+def check_implementation(ref, rtol=2e-02, atol=1e-03)
